+++ judo-runtime-core-dao-rdbms/src/main/java/hu/blackbelt/judo/runtime/core/dao/rdbms/query/model/join/RdbmsNavigationJoin.java
@@ -24,7 +24,9 @@ import hu.blackbelt.judo.meta.asm.runtime.AsmUtils;
 import hu.blackbelt.judo.meta.query.*;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.executors.StatementExecutor;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.RdbmsBuilder;
+import hu.blackbelt.judo.runtime.core.dao.rdbms.query.RdbmsBuilderContext;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.*;
+import hu.blackbelt.judo.runtime.core.dao.rdbms.query.processor.JoinProcessParameters;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.utils.RdbmsAliasUtil;
 import hu.blackbelt.mapper.api.Coercer;
 import lombok.*;
@@ -67,7 +69,7 @@ public class RdbmsNavigationJoin<ID> extends RdbmsJoin {
         super();
         this.query = query;
 
-        final RdbmsBuilder.RdbmsBuilderContext builderContext = RdbmsBuilder.RdbmsBuilderContext.builder()
+        final RdbmsBuilderContext builderContext = RdbmsBuilderContext.builder()
                 .rdbmsBuilder(rdbmsBuilder)
                 .ancestors(subAncestors)
                 .descendants(subDescendants)
@@ -103,7 +105,7 @@ public class RdbmsNavigationJoin<ID> extends RdbmsJoin {
         // list of JOINs using by nested query for embedding to container query
         final List<RdbmsJoin> navigationJoins = navigationJoinList.stream()
                 .flatMap(join -> rdbmsBuilder.processJoin(
-                        RdbmsBuilder.ProcessJoinParameters.builder()
+                        JoinProcessParameters.builder()
                                 .join(join)
                                 .builderContext(builderContext)
                                 .withoutFeatures(withoutFeatures)
@@ -140,7 +142,7 @@ public class RdbmsNavigationJoin<ID> extends RdbmsJoin {
                 additionalJoins.addAll(orderBy.getJoins().stream().flatMap(j -> j.getAllJoins().stream()).collect(Collectors.toList()));
                 for (Join j : additionalJoins) {
                     subJoins.addAll(rdbmsBuilder.processJoin(
-                            RdbmsBuilder.ProcessJoinParameters.builder()
+                            JoinProcessParameters.builder()
                                     .join(j)
                                     .builderContext(builderContext)
                                     .withoutFeatures(withoutFeatures)
